Extract ChildControl to separate implementation file
