+++ CMakeLists.txt
@@ -23,6 +23,8 @@ add_executable(miriway-shell
     miriway_commands.cpp miriway_commands.h
     miriway_policy.cpp   miriway_policy.h
     miriway_workspace_manager.cpp miriway_workspace_manager.h
+        miriway_child_control.cpp
+        miriway_child_control.h
 )
 
 target_include_directories(miriway-shell PUBLIC SYSTEM ${MIRAL_INCLUDE_DIRS})
